+++ lib/chef/knife/helpers/base_vsphere_command.rb
