+++ lib/chef/knife/helpers/customization_helper.rb
