+++ lib/chef/knife/helpers/search_helper.rb
