+++ lib/chef/knife/vsphere_datastore_list.rb
@@ -16,7 +16,7 @@
 # PERFORMANCE OF THIS SOFTWARE
 
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Lists all known data stores in datacenter with sizes
 # VsphereDatastorelist extends the BaseVspherecommand
