+++ lib/chef/knife/vsphere_pool_list.rb
@@ -3,7 +3,7 @@
 # License:: Apache License, Version 2.0
 #
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Lists all known pools in the configured datacenter
 class Chef::Knife::VspherePoolList < Chef::Knife::BaseVsphereCommand
