+++ lib/chef/knife/vsphere_template_list.rb
@@ -4,7 +4,7 @@
 #
 
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Lists all known VM templates in the configured datacenter
 # VsphereTemplatelist extends the BaseVspherecommand
@@ -13,7 +13,7 @@ class Chef::Knife::VsphereTemplateList < Chef::Knife::BaseVsphereCommand
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
