+++ lib/chef/knife/vsphere_vlan_list.rb
@@ -15,7 +15,7 @@
 # PERFORMANCE OF THIS SOFTWARE
 
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Lists all known data stores in datacenter with sizes
 # VsphereVlanlist extends the BaseVspherecommand
