+++ lib/chef/knife/vsphere_vm_move.rb
@@ -3,7 +3,7 @@
 # License:: Apache License, Version 2.0
 #
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Moves the VM to another folder or datastore
 # VsphereVmMove extends the BaseVspherecommand
@@ -12,7 +12,7 @@ class Chef::Knife::VsphereVmMove < Chef::Knife::BaseVsphereCommand
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
