+++ lib/chef/knife/vsphere_vm_network_delete.rb
@@ -3,14 +3,14 @@
 # License:: Apache License, Version 2.0
 #
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 class Chef::Knife::VsphereVmNetworkDelete < Chef::Knife::BaseVsphereCommand
   banner "knife vsphere vm network delete VMNAME NICNAME"
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
