+++ lib/chef/knife/vsphere_vm_network_list.rb
@@ -3,7 +3,7 @@
 # License:: Apache License, Version 2.0
 #
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # VsphereVmNetworklist extends the BaseVspherecommand
 class Chef::Knife::VsphereVmNetworkList < Chef::Knife::BaseVsphereCommand
@@ -11,7 +11,7 @@ class Chef::Knife::VsphereVmNetworkList < Chef::Knife::BaseVsphereCommand
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
