+++ lib/chef/knife/vsphere_vm_show.rb
@@ -2,7 +2,7 @@
 # License:: Apache License, Version 2.0
 
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # VsphereVmShow extends the BaseVspherecommand
 class Chef::Knife::VsphereVmShow < Chef::Knife::BaseVsphereCommand
@@ -10,7 +10,7 @@ class Chef::Knife::VsphereVmShow < Chef::Knife::BaseVsphereCommand
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
