+++ lib/chef/knife/vsphere_vm_state.rb
@@ -4,7 +4,7 @@
 #
 
 require "chef/knife"
-require_relative "base_vsphere_command"
+require_relative "helpers/base_vsphere_command"
 
 # Manage power state of a virtual machine
 # VsphereVmState extends the BaseVspherecommand
@@ -20,7 +20,7 @@ class Chef::Knife::VsphereVmState < Chef::Knife::BaseVsphereCommand
 
   deps do
     Chef::Knife::BaseVsphereCommand.load_deps
-    require_relative "search_helper"
+    require_relative "helpers/search_helper"
     include SearchHelper
   end
